this is right

+++ rmdb/src/storage/disk_manager.cpp
@@ -49,8 +49,9 @@ void DiskManager::read_page(int fd, page_id_t page_no, char *offset, int num_byt
     // 1.lseek()定位到文件头，通过(fd,page_no)可以定位指定页面及其在磁盘文件中的偏移量
     // 2.调用read()函数
     // 注意read返回值与num_bytes不等时，throw InternalError("DiskManager::read_page Error");
-	lseek(fd,page_no*(PAGE_SIZE-1),SEEK_SET);
-	if(read(fd,offset,num_bytes)==-1){
+	lseek(fd,page_no*PAGE_SIZE,SEEK_SET);
+	int x = read(fd,offset,num_bytes);
+	if(x==-1||x!=num_bytes){
 		throw InternalError("DiskManager::read_page Error");
 	}
 }
@@ -129,6 +130,10 @@ void DiskManager::destroy_file(const std::string &path) {
     // 调用unlink()函数
     // 注意不能删除未关闭的文件
 	if(!path2fd_.count(path)){  //path2fd为一个哈希表，.count操作用来查找出现次数，为0说明未打开
+		//这里！！！ 它可能不是存在！！
+		if (!is_file(path)) {
+        	throw FileNotFoundError(path);
+    	}
 		if(unlink(path.c_str())){
 			throw FileNotClosedError(path);
 		}
